(feat) support framebufferinfo
this adds resolution and consumption state into the SMR

flow:
RL will check if front or back is consumed, and ready if possible.
Unreal will check if front or back is ready, and consume if possible.

this ensures there are no race conditions.

framebuffer data will be packed here too eventually